majority-vote/requirements.txt
python_dotenv==1.0.1
